all: unconvert -apply ./...
See https://github.com/mdempsky/unconvert.
Apply it everywhere except the vendored golangorgx code.

Reviewed-on: https://review.gerrithub.io/c/cue-lang/cue/+/1210505
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Unity-Result: CUE porcuepine <cue.porcuepine@gmail.com>

+++ cue/types_test.go
@@ -421,7 +421,7 @@ func TestInt(t *testing.T) {
 
 			vu, err := n.Uint64()
 			checkErr(t, err, tc.errU, "Uint64")
-			if vu != uint64(tc.uint) {
+			if vu != tc.uint {
 				t.Errorf("Uint64: got %v; want %v", vu, tc.uint)
 			}
 		})
